@@ -1,3 +1,3 @@
 module github.com/gravestench/eventemitter
 
-go 1.16
+go 1.19
