@@ -1,227 +1,96 @@
 package eventemitter
 
 import (
-	"fmt"
 	"sync"
 )
 
-// New initializes and returns an EventEmitter instance
-func New() *EventEmitter {
-	ee := &EventEmitter{
-		listeners: make(map[string][]*EventListener),
-		count:     0,
-	}
-
-	return ee
-}
-
+// EventEmitter is a struct that implements the IEventEmitter interface.
 type EventEmitter struct {
-	listeners map[string][]*EventListener
-	count     int
-	mux       sync.Mutex
+	mu            sync.Mutex
+	listeners     map[string][]func(...any)
+	onceListeners map[string][]func(...any)
 }
 
-func (ee *EventEmitter) Emit(e interface{}, args ...interface{}) {
-	var event string
-
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return
-	}
-
-	listeners := ee.listeners[event]
-
-	if listeners == nil {
-		return
-	}
-
-	for idx := range listeners {
-		if listeners[idx].fn != nil {
-			listeners[idx].fn(args...)
-		}
-
-		if listeners[idx].once {
-			listeners = append(listeners[:idx], listeners[idx+1:]...)
-		}
+// New creates a new EventEmitter instance.
+func New() *EventEmitter {
+	return &EventEmitter{
+		listeners:     make(map[string][]func(...any)),
+		onceListeners: make(map[string][]func(...any)),
 	}
 }
 
-func (ee *EventEmitter) On(e interface{}, fn func(...interface{})) {
+// Emit emits an event to all registered listeners.
+func (e *EventEmitter) Emit(event string, args ...any) *sync.WaitGroup {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 
-	var event string
+	var wg sync.WaitGroup
 
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return
+	listeners := e.listeners[event]
+	funcs := make([]func(args ...any), 0)
+	for _, listener := range listeners {
+		wg.Add(1)
+		funcs = append(funcs, func(args ...any) {
+			listener(args...)
+			wg.Done()
+		})
 	}
 
-	ee.addListener(event, fn, false)
+	// Call once listeners and remove them after execution
+	onceListeners := e.onceListeners[event]
+	onceFuncs := make([]func(args ...any), 0)
+	for _, listener := range onceListeners {
+		wg.Add(1)
+		onceFuncs = append(onceFuncs, func(args ...any) {
+			listener(args...)
+			wg.Done()
+		})
+	}
+
+	for _, fn := range funcs {
+		go fn(args...)
+	}
+
+	for _, fn := range onceFuncs {
+		go fn(args...)
+	}
+
+	delete(e.onceListeners, event)
+
+	return &wg
 }
 
-func (ee *EventEmitter) Off(e interface{}, fn func(...interface{})) {
+// On registers a listener for a specific event.
+func (e *EventEmitter) On(event string, fn func(...any)) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 
-	var event string
-
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return
-	}
-
-	ee.removeListener(event, fn)
+	e.listeners[event] = append(e.listeners[event], fn)
 }
 
-func (ee *EventEmitter) Once(e interface{}, fn func(...interface{})) {
+// Off removes a specific listener for a specific event.
+func (e *EventEmitter) Off(event string) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 
-	var event string
-
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return
-	}
-
-	ee.addListener(event, fn, true)
+	delete(e.listeners, event)
 }
 
-func (ee *EventEmitter) addListener(e interface{}, fn func(...interface{}), once bool) *EventEmitter {
-	ee.mux.Lock()
-	defer ee.mux.Unlock()
-	
-	var event string
+// Once registers a one-time listener for a specific event.
+func (e *EventEmitter) Once(event string, fn func(...any)) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return ee
-	}
-
-	if fn == nil {
-		return ee
-	}
-
-	listener := &EventListener{fn, once}
-
-	if ee.listeners[event] == nil {
-		ee.listeners[event] = []*EventListener{listener}
-	} else {
-		ee.listeners[event] = append(ee.listeners[event], listener)
-	}
-
-	return ee
+	e.onceListeners[event] = append(e.onceListeners[event], fn)
 }
 
-func (ee *EventEmitter) removeListener(e interface{}, fn func(...interface{})) {
-	ee.mux.Lock()
-	defer ee.mux.Unlock()
+// RemoveAllListeners removes all listeners for specific event.
+func (e *EventEmitter) RemoveAllListeners(events ...string) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 
-	var event string
-
-	switch s := e.(type) {
-	case string:
-		event = s
-	case fmt.Stringer:
-		event = s.String()
-	default:
-		return
+	for _, event := range events {
+		delete(e.listeners, event)
+		delete(e.onceListeners, event)
 	}
-
-	listeners := ee.listeners[event]
-
-	if listeners == nil {
-		return
-	}
-
-	for idx := range listeners {
-		listenerFn := &listeners[idx].fn
-		removeFn := &fn
-		if listenerFn == removeFn {
-			ee.listeners[event] = append(listeners[:idx], listeners[idx+1:]...)
-		}
-	}
-}
-
-// nolint:unused // not used by anything within the package
-func (ee *EventEmitter) eventNames() []string {
-	names := make([]string, len(ee.listeners))
-
-	idx := 0
-
-	for event := range ee.listeners {
-		names[idx] = event
-		idx++
-	}
-
-	return names
-}
-
-// nolint:unused // not used by anything within the package
-func (ee *EventEmitter) clearEvent(event string) {
-	ee.count--
-	if ee.count <= 0 {
-		ee.count = 0
-		ee.listeners = make(map[string][]*EventListener)
-
-		return
-	}
-
-	delete(ee.listeners, event)
-}
-
-// nolint:unused // not used by anything within the package
-func (ee *EventEmitter) getHandlers(event string) []func(...interface{}) {
-	handlers := make([]func(...interface{}), 0)
-	listeners := ee.listeners[event]
-
-	for idx := range listeners {
-		handlers = append(handlers, listeners[idx].fn)
-	}
-
-	return handlers
-}
-
-// nolint:unused // not used by anything within the package
-func (ee *EventEmitter) getHandlerCount(event string) int {
-	if ee.listeners == nil {
-		return 0
-	}
-
-	listeners := ee.listeners[event]
-	if listeners == nil {
-		return 0
-	}
-
-	return len(listeners)
-}
-
-// nolint:unused // not used by anything within the package
-func (ee *EventEmitter) RemoveAllListeners(events ...string) {
-	if events != nil {
-		if len(events) > 0 {
-			for idx := range events {
-				ee.clearEvent(events[idx])
-			}
-		}
-
-		return
-	}
-
-	ee.listeners = make(map[string][]*EventListener)
 }
